requirements.txt
# requirements.txt

# Core Framework
fastapi
uvicorn[standard]

# Database (SQLAlchemy ORM + PostgreSQL Driver)
sqlalchemy
psycopg2-binary

# Data Validation
pydantic[email]

# Authentication & Hashing
passlib[bcrypt]
python-jose[cryptography]

# Testing
pytest
requests